+++ packages/app/src/cli/services/env/pull.ts
@@ -1,24 +1,18 @@
-import {selectApp} from './app/select-app.js'
-import {AppInterface} from '../models/app/app.js'
+import {selectApp} from '../app/select-app.js'
+import {AppInterface} from '../../models/app/app.js'
 import {output, file} from '@shopify/cli-kit'
 import {patchEnvFile} from '@shopify/cli-kit/node/dot-env'
 import {diffLines} from 'diff'
 
-type Format = 'json' | 'text'
-interface WebEnvOptions {
-  update: boolean
+interface PullEnvOptions {
   envFile: string
 }
 
-export async function webEnv(app: AppInterface, {update, envFile}: WebEnvOptions): Promise<output.Message> {
-  if (update) {
-    return updateEnvFile(app, envFile)
-  } else {
-    return outputEnv(app, 'text')
-  }
+export async function pullEnv(app: AppInterface, {envFile}: PullEnvOptions): Promise<output.Message> {
+  return updateEnvFile(app, envFile)
 }
 
-export async function updateEnvFile(app: AppInterface, envFile: WebEnvOptions['envFile']): Promise<output.Message> {
+export async function updateEnvFile(app: AppInterface, envFile: PullEnvOptions['envFile']): Promise<output.Message> {
   const selectedApp = await selectApp()
 
   const updatedValues = {
@@ -55,23 +49,5 @@ ${output.token.linesDiff(diff)}
 
 ${newEnvFileContent}
 `
-  }
-}
-
-export async function outputEnv(app: AppInterface, format: Format): Promise<output.Message> {
-  const selectedApp = await selectApp()
-
-  if (format === 'json') {
-    return output.content`${output.token.json({
-      SHOPIFY_API_KEY: selectedApp.apiKey,
-      SHOPIFY_API_SECRET: selectedApp.apiSecretKeys[0]?.secret,
-      SCOPES: app.configuration.scopes,
-    })}`
-  } else {
-    return output.content`
-    ${output.token.green('SHOPIFY_API_KEY')}=${selectedApp.apiKey}
-    ${output.token.green('SHOPIFY_API_SECRET')}=${selectedApp.apiSecretKeys[0]?.secret ?? ''}
-    ${output.token.green('SCOPES')}=${app.configuration.scopes}
-  `
   }
 }