+++ packages/app/src/cli/services/environment.test.ts
@@ -1,20 +1,34 @@
-import {fetchAppFromApiKey, fetchOrgAndApps, fetchOrganizations, fetchStoreByDomain} from './dev/fetch.js'
+import {
+  fetchAppExtensionRegistrations,
+  fetchAppFromApiKey,
+  fetchOrgAndApps,
+  fetchOrganizations,
+  fetchStoreByDomain,
+} from './dev/fetch.js'
 import {selectOrCreateApp} from './dev/select-app.js'
 import {selectStore, convertToTestStoreIfNeeded} from './dev/select-store.js'
 import {ensureDeploymentIdsPresence} from './environment/identifiers.js'
-import {DevEnvironmentOptions, ensureDevEnvironment, ensureDeployEnvironment, DeployAppNotFound} from './environment.js'
+import {
+  DevEnvironmentOptions,
+  ensureDevEnvironment,
+  ensureDeployEnvironment,
+  ensureThemeExtensionDevEnvironment,
+  DeployAppNotFound,
+} from './environment.js'
+import {createExtension} from './dev/create-extension.js'
 import {OrganizationApp, OrganizationStore} from '../models/organization.js'
 import {WebType} from '../models/app/app.js'
 import {updateAppIdentifiers, getAppIdentifiers} from '../models/app/identifiers.js'
 import {UIExtension} from '../models/app/extensions.js'
 import {reuseDevConfigPrompt, selectOrganizationPrompt} from '../prompts/dev.js'
-import {testApp} from '../models/app/app.test-data.js'
+import {testApp, testThemeExtensions} from '../models/app/app.test-data.js'
 import metadata from '../metadata.js'
 import {store, api, outputMocker} from '@shopify/cli-kit'
 import {beforeEach, describe, expect, it, test, vi} from 'vitest'
 
 beforeEach(() => {
   vi.mock('./dev/fetch')
+  vi.mock('./dev/create-extension')
   vi.mock('./dev/select-app')
   vi.mock('./dev/select-store')
   vi.mock('../prompts/dev')
@@ -49,6 +63,7 @@ const APP1: OrganizationApp = {
   apiKey: 'key1',
   organizationId: '1',
   apiSecretKeys: [{secret: 'secret1'}],
+  grantedScopes: [],
 }
 const APP2: OrganizationApp = {
   id: '2',
@@ -56,6 +71,7 @@ const APP2: OrganizationApp = {
   apiKey: 'key2',
   organizationId: '1',
   apiSecretKeys: [{secret: 'secret2'}],
+  grantedScopes: [],
 }
 
 const ORG1: api.graphql.AllOrganizationsQuerySchemaOrganization = {
@@ -105,7 +121,7 @@ const LOCAL_APP = testApp({
   name: 'my-app',
   directory: '/app',
   configurationPath: '/shopify.app.toml',
-  configuration: {scopes: 'read_products'},
+  configuration: {scopes: 'read_products', extensionDirectories: ['extensions/*']},
   webs: [
     {
       directory: '',
@@ -475,5 +491,68 @@ describe('ensureDeployEnvironment', () => {
     expect(got.partnersApp.title).toEqual(APP1.title)
     expect(got.partnersApp.appType).toEqual(APP1.appType)
     expect(got.identifiers).toEqual({app: APP1.apiKey, extensions: {}, extensionIds: {}})
+  })
+})
+
+describe('ensureThemeExtensionDevEnvironment', () => {
+  test('fetches theme extension when it exists', async () => {
+    // Given
+    const token = 'token'
+    const apiKey = 'apiKey'
+    const extension = testThemeExtensions()
+
+    vi.mocked(fetchAppExtensionRegistrations).mockResolvedValue({
+      app: {
+        extensionRegistrations: [
+          {
+            id: 'other ID',
+            uuid: 'other UUID',
+            title: 'other extension',
+            type: 'other',
+          },
+          {
+            id: 'existing ID',
+            uuid: 'UUID',
+            title: 'theme app extension',
+            type: 'THEME_APP_EXTENSION',
+          },
+        ],
+      },
+    })
+
+    // When
+    const got = await ensureThemeExtensionDevEnvironment(extension, apiKey, token)
+
+    // Then
+    expect('existing ID').toEqual(got.id)
+    expect('UUID').toEqual(got.uuid)
+    expect('theme app extension').toEqual(got.title)
+    expect('THEME_APP_EXTENSION').toEqual(got.type)
+  })
+
+  test('creates theme extension when it does not exist', async () => {
+    // Given
+    const token = 'token'
+    const apiKey = 'apiKey'
+    const extension = testThemeExtensions()
+
+    vi.mocked(fetchAppExtensionRegistrations).mockResolvedValue({
+      app: {extensionRegistrations: []},
+    })
+    vi.mocked(createExtension).mockResolvedValue({
+      id: 'new ID',
+      uuid: 'UUID',
+      title: 'theme app extension',
+      type: 'THEME_APP_EXTENSION',
+    })
+
+    // When
+    const got = await ensureThemeExtensionDevEnvironment(extension, apiKey, token)
+
+    // Then
+    expect('new ID').toEqual(got.id)
+    expect('UUID').toEqual(got.uuid)
+    expect('theme app extension').toEqual(got.title)
+    expect('THEME_APP_EXTENSION').toEqual(got.type)
   })
 })