+++ packages/cli-hydrogen/src/cli/services/deploy/error.ts
@@ -1,0 +1,5 @@
+import {error} from '@shopify/cli-kit'
+
+export const WebPageNotAvailable = () => {
+  return new error.Abort('Web page not available.')
+}