+++ packages/cli-main/bin/dev.js
@@ -2,8 +2,6 @@
 
 process.removeAllListeners('warning');
 
-process.env.SHOPIFY_ENV = process.env.SHOPIFY_ENV ?? "development"
-
 import runCLI from "../dist/index.js";
 
-runCLI();
+runCLI({development: true});